feat: adds helpers directory
