+++ client/src/helpers/decodeHtmlEntities.ts
@@ -1,0 +1,4 @@
+export const decodeHtmlEntities = (html: string) => {
+  const doc = new DOMParser().parseFromString(html, "text/html");
+  return doc.body.textContent || "";
+};
